Jamie

+++ robot_code/system_demo/RobotHand.h
@@ -32,8 +32,8 @@ class RobotHand
   public:
     RobotHand();
     void init(int left_step_pin, int left_direction_pin, int right_step_pin, int right_direction_pin, int pinch_pin_1, int pinch_pin_2);
-    float remapAngle(float old_angle, float old_min, float old_max, float new_min, float new_max);
-    float remap(float old_val, float old_min, float old_max, float new_min, float new_max);
+    int remapAngle(float old_angle, float old_min, float old_max, float new_min, float new_max);
+    int remap(float old_val, float old_min, float old_max, float new_min, float new_max);
     void setOrientation(int pitch_angle, int roll_angle);
     void setClaw(int pointerStatus, int thumbStatus);
     void updateSensors();
